% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gl_geocode.R
\name{gl_geocode}
\alias{gl_geocode}
\title{Geolocation via Google Translate}
\usage{
gl_geocode(str, region, key)
}
\arguments{
\item{str}{vector of strings}

\item{region}{bias results towards a region}

\item{key}{API key}
}
\value{
A list of \code{tibbles} with list columns.
}
\description{
Slick function to geolocate via the Google Maps API
}
\examples{

library(tidyverse)

gl_geocode(c("Konstanz", "Mannheim", "Köln"), key=your_api_key, region='DE')

# tidyr 
tibble(str=c("Konstanz", "Mannheim", "Köln")) \%>\% 
	mutate(geo=gl_geocode(str,
	 region='DE', key=api_key)) \%>\% 
 unnest(col='geo')
 
}
