% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gl_place_details.R
\name{gl_place_details}
\alias{gl_place_details}
\title{Google Place ID API}
\usage{
gl_place_details(place_id, key, fields)
}
\arguments{
\item{place_id}{Google Place ID}

\item{key}{API key}

\item{fields}{fields to retrieve}
}
\value{
A list.
}
\description{
Slick function to retrieve information from Google Place API
}
\details{
Available fields are listed in the API Dev Guides \url{https://developers.google.com/places/web-service/details}.
}
\examples{

gl_place_details(
 place_id='ChIJ3wZkgmIbkUcRYEgkDb9IjMY', 
 key=your_api_key, fields='formatted_address,geometry')

}
