% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gl_translate.R
\name{gl_translate}
\alias{gl_translate}
\title{Translate via Google Translate}
\usage{
gl_translate(str, target, key)
}
\arguments{
\item{str}{vector of strings}

\item{target}{target language code}

\item{key}{API key}
}
\value{
A list of \code{tibbles}.
}
\description{
Slick function to retrieve translation from the Google Translate API
}
\examples{

library(tidyverse)

gl_translate(c("Bonjour", "hallo", "ola"), target='en', key=your_api_key)

# tidyr 
tibble(str=c("Bonjour", "hallo", "ola")) \%>\% 
	mutate(translated=gl_translate(str,
	 target='en', key=api_key)) \%>\% 
 unnest(col='translated')
 
}
